Garage Door Control System Using the ESP8266
This control system replaces the typical remote controls provided as
standard with modern garage door systems. The ESP8266 with nodeMCU
firmware is used as the core controller. Activation of the system is via
a mechanical relay wired in parallel with the wall mounted door switch.
Garage door opening status is detected via a magnetic switch sensor
typically used in security alarm systems.